backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
passlib[bcrypt]==1.7.4
python-jose==3.5.0
email-validator==2.2.0
scikit-learn==1.7.1
pandas==2.3.2
numpy==2.3.2
scipy==1.16.1
joblib==1.5.2
python-dotenv==1.1.1
